better file linking, icons, backgrounds

+++ setup.sh
@@ -17,8 +17,18 @@ echo "- Linking all config folders"
 for dir in config/*;
 do
     echo "-- Setting up $dir"
-    rm -f ~/.$dir
-    ln -s `pwd`/$dir ~/.$dir
+    if test -f ~/.$dir;
+    then
+        rm -f ~/.$dir
+    fi
+
+    mkdir ~/.$dir 2>/dev/null
+
+    for file in `pwd`/$dir;
+    do
+        rm -f ~/.$file 2>/dev/null || true
+        ln -s `pwd`/$file ~/.$file
+    done
 done
 
 mkdir ~/.zsh 2>/dev/null || true
@@ -47,9 +57,17 @@ fi
 
 CTP_CURSOR_VERSION="v0.2.0"
 if [ ! -e ~/.themes/Catppuccin-Mocha-Dark-Cursors ]; then \
-  mkdir -p ~/.themes \
+  mkdir -p ~/.icons \
   && curl -L https://github.com/catppuccin/cursors/releases/download/$CTP_CURSOR_VERSION/Catppuccin-Mocha-Dark-Cursors.zip -o ~/.themes/catppuccin.zip \
-  && unzip ~/.themes/catppuccin.zip -d ~/.themes/ \
+  && unzip ~/.themes/catppuccin.zip -d ~/.icons/ \
   && rm -rf ~/.themes/catppuccin.zip;
 fi
+
+curl \
+    -L https://raw.githubusercontent.com/NixOS/nixos-artwork/master/wallpapers/nix-wallpaper-nineish-dark-gray.png \
+    -o ~/.config/sway/lockscreen.png
+
+curl \
+    -L https://w.wallhaven.cc/full/m9/wallhaven-m9lxe9.jpg \
+    -o ~/.config/sway/background.jpg
